+++ packages/dynamic_tabs/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 04.06.2019
 
-- Updating Example to be desktop aware
+* Updating Example to be desktop aware
 
 ## 0.3.0 - 04.05.2019